BUG #13393. Enable GMV reader only when OpenGLU is available.
Enable GMV reader only when GLU is available. Also the option to optionally turn
ON GMV reader didn't seem to work. Despite the fact the the GMV reader was
turned on, it wasn't being built causing dashboard failures. This patch
overcomes the issue by not making GMV user-optional, instead simply disabling it
automatically when GLU was not available.