BUG #0014102. Fix CGNS segfaults when cell/point vars are named same.
Temporary fix for BUG #14102. The issue comes from the fact that they are trying
to load celldata and pointdata variables with the same name and the reader can't
sort them ( the location information is not pass through the function call
"GetVar"). I have written a temporary fix that will prevent the crash but not
solve the problem.  In the attached file, the chicken path I take is very
simple: I modify variable name by prepending "CELL_" when cell centered data are
found, then if point data exist it's kept to prevent possible collision,
otherwise it's removed.  In the end, as long as user is not manipulating mixed
Cell and Vertex data, nothing is changed. On the contrary, if there is any risk
of crash CELL data will appear with a prefix.