Making VisIt bridge compatible with the way readers are loaded.
ParaView was changed so that the definition of the readers
(including their extensions) was obtained solely from the server
instead of using a special file to associate a reader with
a file extension on the client. That extension association was
already in most of the XML descriptions in the hints for readers
anyways. So now as long as a reader is built on the server
it will be available on the client, independent of whether the
client was "built" with that reader. These changes to the VisIt
bridge take advantage of those changes.