Revert "Move visit reader stubs from CMake to AvtAlgorithms directory."
This reverts commit 4664bf4024275351b4193e5a3f913d36b7bb764a.

The said commit moved *.in files to AvtAlgorithms, but that raise a
couple of issues:

1. Install rules were broken. The install rules would have needed to be
fixed.
2. The VisItCommon.cmake file which is intended  to be imported in
plugins external to this source repository now referred to t he
CMAKE_CURRENT_SOURCE_DIR to locate AvtAlgorithms dir. That too was
incorrect and would have needed to be fixed and tested.

Instead opting to revert the change since there was no strong reason to
move the *.in files in the first place.

Conflicts:
	databases/CMakeLists.txt