Correct Nek5000 reading when on windows.
The Nek5000 reader presumes that when parsing a file by default
it will only encounter '\n' newline characters. But by default
opening a file actually means to use that system new line character(s).
So on windows all '\n' are converted to '\r\n'. To solve this problem
we explicitly open all files in binary mode which no conversion of line
endings.