Converted internal proxies from SourceProxy to Proxy.
When internal (sub) proxies are classified as SourceProxy,
the GUI forces things like UpdateInformation() on them which
can lead to subtle bugs, especially if they are part of a
vtkFileSeriesReader. Classifying them as Proxy fixes this issue

Also, added series extensions to the Silo reader to enable
the new meta format. More of those changes later.